requirements.txt
googlemaps==4.10.0
langchain==0.0.263
openai==0.27.8
python-dotenv==1.0.0
google-generativeai==0.1.0rc1
leafmap==0.23.4
geopandas==0.13.2
geopy
folium
streamlit
gradio
branca==0.6.0
streamlit_folium
